requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==20.3.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clean-text==0.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==5.2.4
django-appconf==1.2.0
django_compressor==4.6.0
emoji==1.7.0
exceptiongroup==1.3.1
fonttools==4.61.1
fpdf==1.7.2
ftfy==6.3.1
idna==3.11
jellyfish==1.2.1
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.1
language_tool_python==3.2.2
lib50==3.1.4
Markdown==3.8.2
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pexpect==4.9.0
pillow==12.1.0
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pycparser==2.23
pyparsing==3.3.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rcssmin==1.2.2
redis==7.1.0
requests==2.32.5
rjsmin==1.2.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
submit50==3.2.1
termcolor==1.1.0
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wordcloud==1.9.5
